--- src/main/java/com/isa/unasdziala/domain/Employee.java
@@ -1,27 +1,0 @@
-package com.isa.unasdziala.domain;
-
-import com.isa.unasdziala.domain.converter.DepartmentConverter;
-import com.opencsv.bean.CsvBindByPosition;
-import com.opencsv.bean.CsvCustomBindByPosition;
-import com.opencsv.bean.CsvRecurse;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Employee {
-    @CsvBindByPosition(position = 0)
-    private String firstName;
-    @CsvBindByPosition(position = 1)
-    private String lastName;
-    @CsvRecurse
-    private Contact contact;
-    @CsvRecurse
-    private Address address;
-    @CsvCustomBindByPosition(position = 10, converter = DepartmentConverter.class)
-    private Department department;
-    @CsvBindByPosition(position = 11)
-    private float holidays;
-}
