+++ src/main/java/com/isa/unasdziala/utils/CalendarLoader.java
@@ -1,7 +1,6 @@
 package com.isa.unasdziala.utils;
 
 import com.isa.unasdziala.domain.Day;
-import com.isa.unasdziala.domain.Employee;
 import net.fortuna.ical4j.data.CalendarBuilder;
 import net.fortuna.ical4j.data.ParserException;
 import net.fortuna.ical4j.model.Calendar;
