+++ src/test/java/com/isa/unasdziala/utils/CalendarLoaderTest.java
@@ -1,7 +1,6 @@
 package com.isa.unasdziala.utils;
 
 import com.isa.unasdziala.domain.Day;
-import com.isa.unasdziala.domain.Employee;
 import org.junit.jupiter.api.Test;
 
 import java.util.Set;
